Split common tests and tpc-ds stress test on different modules
And add option to run TPC-DS stress test separately.
